Added a --tar-archive-dir option to save tar files (optionally) between runs.

+++ superflore/generators/bitbake/yocto_recipe.py
@@ -23,17 +23,18 @@
 # IN THE SOFTWARE.
 #
 
-import errno
 import hashlib
-import os
+import os.path
 import sys
 import tarfile
 
+from rosinstall_generator.distro import get_distro
+
 from superflore.exceptions import NoPkgXml
 
+from superflore.utils import get_pkg_version
 from superflore.utils import info
 from superflore.utils import resolve_dep
-from superflore.utils import warn
 
 if sys.version_info[0] == 2:
     import requests
@@ -51,10 +52,10 @@ else:
 
 
 class yoctoRecipe(object):
-    def __init__(self, name, distro, src_uri):
+    def __init__(self, name, distro, src_uri, tar_dir):
         self.name = name
-        self.distro = distro
-        self.version = None
+        self.distro = distro.name
+        self.version = get_pkg_version(distro, name)
         self.description = ''
         self.src_uri = src_uri
         self.pkg_xml = None
@@ -62,40 +63,24 @@ class yoctoRecipe(object):
         self.license = None
         self.depends = list()
         self.license_line = None
+        self.archive_name = None
         self.license_md5 = None
-        self.src_md5 = None
-        self.src_sha256 = None
-
-    def __enter__(self):
+        self.tar_dir = tar_dir
         self.downloadArchive()
         self.src_sha256 = hashlib.sha256(
-            open("./" + self.getArchiveName(), 'rb').read()).hexdigest()
+            open(self.getArchiveName(), 'rb').read()).hexdigest()
         self.src_md5 = hashlib.md5(
-            open("./" + self.getArchiveName(), 'rb').read()).hexdigest()
-        return self
-
-    def __exit__(self, *args):
-        info(
-            "removing downloaded package archive '{0}'...".format(
-                self.getArchiveName()
-            )
-        )
-        try:
-            os.remove(self.getArchiveName())
-        except OSError as oe:
-            if oe.errno == errno.ENOENT:
-                warn("file '%s' not found!" % self.getArchiveName())
-                pass
-            else:
-                raise
+            open(self.getArchiveName(), 'rb').read()).hexdigest()
 
     def getFolderName(self):
         return self.name.replace("-", "_") + "-" + str(self.version)
 
     def getArchiveName(self):
-        return self.name.replace("-", "_") + \
-               "-" + str(self.version) + \
-               ".tar.gz"
+        if not self.archive_name:
+            self.archive_name = self.tar_dir + "/" \
+                + self.name.replace('-', '_') + '-' + str(self.version) \
+                + '-' + self.distro + '.tar.gz'
+        return self.archive_name
 
     def get_license_line(self):
         self.license_line = ''
@@ -113,8 +98,11 @@ class yoctoRecipe(object):
                 break
 
     def downloadArchive(self):
-        info("downloading archive version for package '%s'..." % self.name)
-        urllib.request.urlretrieve(self.src_uri, self.getArchiveName())
+        if os.path.exists(self.getArchiveName()):
+            info("using cached archive for package '%s'..." % self.name)
+        else:
+            info("downloading archive version for package '%s'..." % self.name)
+            urllib.request.urlretrieve(self.src_uri, self.getArchiveName())
 
     def extractArchive(self):
         tar = tarfile.open(self.getArchiveName(), "r:gz")
